LayoutTests/imported/w3c/web-platform-tests/resize-observer/eventloop-expected.txt

CONSOLE MESSAGE: ResizeObserver loop completed with undelivered notifications.
ResizeObserver notification event loop tests

t1

Harness Error (TIMEOUT), message = null

PASS ResizeObserver implemented 
NOTRUN guard 
FAIL test0: multiple notifications inside same event loop assert_equals: new loop expected 1 but got 0
